+++ src/commands/music/disconnect.rs
@@ -9,7 +9,7 @@ use crate::bot::Context;
 #[poise::command(slash_command, guild_only)]
 pub async fn disconnect(ctx: Context<'_>) -> Result<(), Error> {
     let manager = ctx.data();
-    let guild = ctx.guild().expect("poise lied to me").id;
+    let guild = ctx.guild_id().expect("poise lied to me");
 
     let Some(session) = manager.get_session(SessionQuery::Guild(guild)) else {
         ctx.send(