+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
-## 2.2.0 | August 8th 2024
+## 2.2.1 | August 22nd 2024
+
+- Added new option: `/playing` can now receive an updating behavior parameter
+- Added album name to `/playing` embed
+- Fixed a bug where uncached guilds would panic the bot
+- Fixed small issue with embed styling
+- Updated to Rust 1.80.1 (from 1.79.0)
+- Updated `diesel` and addons to latest versions
+- Removed `lazy_static` in favor of `LazyLock` (Rust 1.80.0+ feature)
+- Bumped MSRV to 1.80.0 due to the introduction of `LazyLock`
+
+## 2.2.0 | August 13th 2024
 
 ### Changes